# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "My Store",
      "description": "Juan Perez is a very popular guy here in my country. So I decided to create an account: id=jperez password=123456. The account has the 'guest' role which means than. Once you login with that id and password I recommend don't refresh the page. If you do, you'll have to login one more time",
      "contact": {
        "name": "Sandro Simón",
        "email": "hello@sandrosimon.com"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "/api/v1",
        "description": "Server"
      }
      
    ],
...
    "paths": {
...
      "/users": {
        "get": {
          "tags": ["User"],
          "summary": "User roles with access to this endpoint: admin",
          "description": "Get all users",
          "operationId": "getAllUsers",
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref":"#/components/schemas/User"
                    }  
                  },
                  "example": [{
                    "id": "ssimon",
                    "role": "admin",
                    "email": "hello@sandrosimon.com",
                    "firstName": "SANDRO",
                    "lastName": "SIMON"
                  }]
                }
              }
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "post": {
          "tags": ["User"],
          "summary": "User roles with access to this endpoint: admin",
          "description": "Add a new user to the database",
          "operationId": "createOneUser",
          "requestBody": {
            "description": "Create a new user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/User"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/User"  
                  }
                }
                
              }
            },
            "405": {
              "description": "Invalid input"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }      
      },
      "/users/{id}": {
        "get": {
          "tags": ["User"],
          "summary": "User roles with access to this endpoint: admin, seller, guest",
          "description": "Returns an user",
          "operationId": "getOneUser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Code of the account to return",
              "required": true,
              "schema": {
                "type": "string",
                "example": "jperez"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref":"#/components/schemas/User"
                  },
                  "example": {
                    "id": "ssimon",
                    "role": "admin",
                    "email": "hello@sandrosimon.com",
                    "firstName": "SANDRO",
                    "lastName": "SIMON"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Account not found"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        },
        "put": {
          "tags": ["User"],
          "summary": "User roles with access to this endpoint: admin",
          "description": "Update an existing user by Id",
          "operationId": "updateOneUser",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Code of the user to return",
              "required": true,
              "schema": {
                "type": "string",
                "example": "ssimon"
              }
            } 
          ],
          "requestBody": {
            "description": "Update an existent account in the store",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                },
                "example": {
                  "role": "admin",
                  "password": "1234asd",
                  "email": "hello@sandrosimon.com",
                  "firstName": "SANDRO",
                  "lastName": "SIMON"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid account code supplied"
            },
            "404": {
              "description": "Account not found"
            },
            "405": {
              "description": "Validation exception"
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
          
        },
        "delete": {
          "tags": ["User"],
          "summary": "User roles with access to this endpoint: admin",
          "description": "delete an account",
          "operationId": "deleteOneAccount",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Account code to delete",
              "required": true,
              "schema": {
                "type": "string",
                "example": "jperez"
              }
            }
          ],
          "responses": {
            "204":{
              "description": "The account was deleted successfully."
            },
            "500":{
              "description": "Unexpected error."
            }
          },
          "security": [
            { "storeAuth": [] }
          ]
        }
      }  
    },
    "components": {
      "schemas": {
...
        "User": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "lgalindo"
            },
            "role": {
              "type": "string",
              "example": "admin"
            },
            "password": {
              "type": "string",
              "example": "admin123"
            },
            "email": {
              "type": "string",
              "example": "admin@database.com"
            },
            "firstName": {
              "type": "string",
              "example": "Lizbeth"
            },
            "lastName": {
              "type": "string",
              "example": "Galindo"
            }
          }
        },
...
      },
      "securitySchemes": {
        "storeAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }  
  }